+++ src/main/java/com/auction/ecommerce/repository/AuctionRepository.java
@@ -3,5 +3,5 @@ package com.auction.ecommerce.repository;
 import com.auction.ecommerce.model.Auction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuctionRepository extends JpaRepository<Auction, Long> {
+public interface AuctionRepository extends JpaRepository<Auction, Integer> {
 }
